+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+###0.4.7-pre / 2016-12-19
+* the Graphite and text file reporters now work with configurable error tolerance
+
+###0.4.6-pre / 2016-12-06
+* reporting can now be configured to be error tolerant (Metrics.Reports.ToleratedConsecutiveFailures config key)
+* an exception is now thrown during startup if an http endpoint is configured more than once
+
+###0.4.5-pre / 2016-12-02
+* disabling metrics now properly works with config extensions
+* performance counters are no longer being registered when metrics are disabled
+
 ###0.4.4 / 2016-12-02
 * health checks can now be updated at runtime (@bronsh)
 * fixed the 404 response of core metrics http listener